+++ src/main/java/blackjack/domain/card/Cards.java
@@ -7,6 +7,8 @@ import java.util.Set;
 
 public class Cards {
 
+    private static final int ACE_ADDITIONAL_NUMBER = 10;
+    private static final int BEST_SCORE = 21;
     private static final String DUPLICATE_EXCEPTION_MESSAGE = "카드 패에 중복된 카드가 존재할 수 없습니다.";
 
     private final List<Card> cardHand;
@@ -34,5 +36,25 @@ public class Cards {
 
     public int getSize() {
         return cardHand.size();
+    }
+
+    public int getBestScore() {
+        int sum = cardHand.stream()
+                .map(Card::getNumber)
+                .map(Number::getScore)
+                .reduce(0, Integer::sum);
+
+        for (Card card : cardHand) {
+            sum = getBest(sum, card);
+        }
+
+        return sum;
+    }
+
+    private int getBest(int sum, Card card) {
+        if (card.isAce() && sum + ACE_ADDITIONAL_NUMBER <= BEST_SCORE) {
+            sum += ACE_ADDITIONAL_NUMBER;
+        }
+        return sum;
     }
 }
